refactor(capabilities): :recycle: relocate shared hooks to design-system/hooks/ (flat home)
Move all 19 hook files from src/components/design-system/utils/hooks/ to the new
flat src/components/design-system/hooks/ directory (the exempt shared-hooks home).
Move the non-hook helper shuffle-array.ts to design-system/utils/shuffle-array.ts.
Update all 50+ import references across 30 files (mix of @/ absolute and relative paths).
Remove the now-empty utils/hooks/ directory.
