+++ src/components/design-system/hooks/use-device-capabilities.ts
