+++ src/components/design-system/hooks/use-shuffle-array.ts
@@ -2,7 +2,7 @@
 'use client';
 
 import { useState, useEffect } from "react";
-import { shuffleArray } from "./shuffle-array";
+import { shuffleArray } from "@/components/design-system/utils/shuffle-array";
 
 export const useShuffleArray = <T>(array: T[] = [], numberOfItems: number): T[] => {
   const [shuffled, setShuffled] = useState<T[]>(array.slice(0, numberOfItems));
