+++ src/components/design-system/organism/image-carousel.tsx
@@ -6,7 +6,7 @@ import { BiChevronLeft, BiChevronRight, BiX } from "react-icons/bi";
 import { ImageGlow } from "../atoms/effects/image-glow";
 import { Button } from "../atoms/buttons/button";
 import { MotionDiv } from "../molecules/animation/motion-div";
-import { useGlassmorphism } from "../utils/hooks/use-glassmorphism";
+import { useGlassmorphism } from "@/components/design-system/hooks/use-glassmorphism";
 import Image from "next/image";
 import { imageShimmerPlaceholder } from "../atoms/effects/image-shimmer-placeholder";
