+++ src/components/design-system/organism/menu.tsx
@@ -12,9 +12,9 @@ import { Close } from "../molecules/menu/close";
 import { DropdownMenu } from "../molecules/menu/dropdown-menu";
 import { HamburgerMenu } from "../molecules/menu/hamburger-menu";
 import { MenuItemWithTracking } from "../molecules/menu/menu-item-with-tracking";
-import { useGlassmorphism } from "../utils/hooks/use-glassmorphism";
-import { ScrollDirection, useScrollDirection } from "../utils/hooks/use-scroll-direction";
-import { useOsModifierKey } from "../utils/hooks/use-os-modifier-key";
+import { useGlassmorphism } from "@/components/design-system/hooks/use-glassmorphism";
+import { ScrollDirection, useScrollDirection } from "@/components/design-system/hooks/use-scroll-direction";
+import { useOsModifierKey } from "@/components/design-system/hooks/use-os-modifier-key";
 import { MotionDiv } from "../molecules/animation/motion-div";
 import { LuCommand } from "react-icons/lu";
 import { ImCtrl } from "react-icons/im";
