+++ src/components/design-system/organism/tracking-optin.tsx
@@ -3,7 +3,7 @@
 import { GoogleAnalytics } from "@next/third-parties/google";
 import { Analytics } from "@vercel/analytics/next";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import { useConsentStore } from "../utils/hooks/use-consent-store";
+import { useConsentStore } from "@/components/design-system/hooks/use-consent-store";
 
 const TrackingOptIn = () => {
   const enabled = useConsentStore();
