+++ src/content/blog/post/2026/04/05/chrome-built-in-ai-summarizer/content.mdx
@@ -115,7 +115,7 @@ Than I create the `useChromeSummarize` hook, the core of the feature. This is a 
 "use client";
 
 import { useCallback, useEffect, useState, useRef } from "react";
-import { useDeviceCapabilities } from "@/components/design-system/utils/hooks/use-device-capabilities";
+import { useDeviceCapabilities } from "@/components/design-system/hooks/use-device-capabilities";
 
 export type SummaryType = "tldr" | "key-points";
 export type SummaryStatus = "idle" | "downloading" | "loading" | "streaming" | "done" | "error";
